Refactor CallViewController environment initialization
